requirements.txt
python-dateutil==2.8.2
pytz==2022.2.1
simple-term-menu==1.5.0
six==1.16.0
